Persiste os dados das pessoas em arquivo texto.

+++ src/entities/Interpessoal.java
@@ -4,6 +4,6 @@ public class Interpessoal implements Habilidade {
 
     @Override
     public int calculaNivel(int autoestima, int nivelEmpatia, int experienciaProfisional) {
-        return 0.3 * autoestima + 0.2 * (experienciaProfisional / 10) + 0.5 * nivelEmpatia;
+        return (int)(0.3 * autoestima + 0.2 * (experienciaProfisional / 10) + 0.5 * nivelEmpatia);
     }
 }
